requirements.txt
aiogram
tortoise-orm
annotated-types==0.7.0
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
typing-extensions==4.13.1
typing-inspection==0.4.0
